renamed from text2sql to premsql
